add download script & add deletion check

+++ demo.py
@@ -43,7 +43,7 @@ def demo(opt):
     frame_rate = dataloader.frame_rate
 
     frame_dir = None if opt.output_format == 'text' else osp.join(result_root, new_name)
-    if frame_dir:
+    if frame_dir and osp.exists(frame_dir):
         shutil.rmtree(frame_dir) 
     eval_seq(opt, dataloader, mot_model, 'mot', result_filename,
              save_dir=frame_dir, show_image=False, frame_rate=frame_rate,
